moving gallery apis to pxtlib
